extracted shared go code
