+++ binding/go/porcupine_unix.go
@@ -77,101 +77,8 @@ import (
 	"unsafe"
 )
 
-//go:embed embedded
-var embeddedFS embed.FS
-
-// PvStatus type
-type PvStatus int
-
-// Possible status return codes from the Porcupine library
-const (
-	SUCCESS          PvStatus = 0
-	OUT_OF_MEMORY    PvStatus = 1
-	IO_ERROR         PvStatus = 2
-	INVALID_ARGUMENT PvStatus = 3
-)
-
-func pvStatusToString(status PvStatus) string {
-	switch status {
-	case SUCCESS:
-		return "SUCCESS"
-	case OUT_OF_MEMORY:
-		return "OUT_OF_MEMORY"
-	case IO_ERROR:
-		return "IO_ERROR"
-	case INVALID_ARGUMENT:
-		return "INVALID_ARGUMENT"
-	default:
-		return "Unknown error"
-	}
-}
-
-// BuiltInKeyword Type
-type BuiltInKeyword string
-
-// Available built-in wake words constants
-const (
-	ALEXA       BuiltInKeyword = "alexa"
-	AMERICANO   BuiltInKeyword = "americano"
-	BLUEBERRY   BuiltInKeyword = "blueberry"
-	BUMBLEBEE   BuiltInKeyword = "bumblebee"
-	COMPUTER    BuiltInKeyword = "computer"
-	GRAPEFRUIT  BuiltInKeyword = "grapefruit"
-	GRASSHOPPER BuiltInKeyword = "grasshopper"
-	HEY_BARISTA BuiltInKeyword = "hey barista"
-	HEY_GOOGLE  BuiltInKeyword = "hey google"
-	HEY_SIRI    BuiltInKeyword = "hey siri"
-	JARVIS      BuiltInKeyword = "jarvis"
-	OK_GOOGLE   BuiltInKeyword = "ok google"
-	PICO_CLOCK  BuiltInKeyword = "pico clock"
-	PICOVOICE   BuiltInKeyword = "picovoice"
-	PORCUPINE   BuiltInKeyword = "porcupine"
-	TERMINATOR  BuiltInKeyword = "terminator"
-)
-
-// List of available built-in wake words
-var BuiltInKeywords = []BuiltInKeyword{
-	ALEXA, AMERICANO, BLUEBERRY, BUMBLEBEE, COMPUTER, GRAPEFRUIT, GRASSHOPPER, HEY_BARISTA,
-	HEY_GOOGLE, HEY_SIRI, JARVIS, OK_GOOGLE, PICO_CLOCK, PICOVOICE, PORCUPINE, TERMINATOR,
-}
-
-// Checks if a given BuiltInKeyword is valid
-func (k BuiltInKeyword) IsValid() error {
-	for _, b := range BuiltInKeywords {
-		if k == b {
-			return nil
-		}
-	}
-	return errors.New("Invalid built-in keyword.")
-}
-
-// Porcupine struct
-type Porcupine struct {
-	handle uintptr
-
-	// Absolute path to the file containing model parameters.
-	ModelPath string
-
-	// Sensitivity values for detecting keywords. Each value should be a number within [0, 1]. A
-	// higher sensitivity results in fewer misses at the cost of increasing the false alarm rate.
-	Sensitivities []float32
-
-	// List of built-in keywords to use.
-	BuiltInKeywords []BuiltInKeyword
-
-	// Absolute paths to keyword model files.
-	KeywordPaths []string
-}
-
 // private vars
 var (
-	osName        = getOS()
-	extractionDir = filepath.Join(os.TempDir(), "porcupine")
-
-	defaultModelFile = extractDefaultModel()
-	builtinKeywords  = extractKeywordFiles()
-	libName          = extractLib()
-
 	lib                           = C.dlopen(C.CString(libName), C.RTLD_NOW)
 	pv_porcupine_init_ptr         = C.dlsym(lib, C.CString("pv_porcupine_init"))
 	pv_porcupine_process_ptr      = C.dlsym(lib, C.CString("pv_porcupine_process"))
@@ -181,55 +88,9 @@ var (
 	pv_porcupine_delete_ptr       = C.dlsym(lib, C.CString("pv_porcupine_delete"))
 )
 
-var (
-	// Number of audio samples per frame.
-	FrameLength = int(C.pv_porcupine_frame_length_wrapper(pv_porcupine_frame_length_ptr))
-
-	// Audio sample rate accepted by Picovoice.
-	SampleRate = int(C.pv_sample_rate_wrapper(pv_sample_rate_ptr))
-
-	// Porcupine version
-	Version = C.GoString(C.pv_porcupine_version_wrapper(pv_porcupine_version_ptr))
-)
-
-// Init function for Porcupine. Must be called before attempting process
-func (porcupine *Porcupine) Init() (err error) {
-	if porcupine.ModelPath == "" {
-		porcupine.ModelPath = defaultModelFile
-	}
-
-	if _, err := os.Stat(porcupine.ModelPath); os.IsNotExist(err) {
-		return fmt.Errorf("%s: Specified model file could not be found at %s", pvStatusToString(INVALID_ARGUMENT), porcupine.ModelPath)
-	}
-
-	if porcupine.BuiltInKeywords != nil && len(porcupine.BuiltInKeywords) > 0 {
-		for _, keyword := range porcupine.BuiltInKeywords {
-			if err := keyword.IsValid(); err != nil {
-				return fmt.Errorf("%s: '%s' is not a valid built-in keyword.", pvStatusToString(INVALID_ARGUMENT), keyword)
-			}
-			keywordStr := string(keyword)
-			porcupine.KeywordPaths = append(porcupine.KeywordPaths, builtinKeywords[keywordStr])
-		}
-	}
-
-	if porcupine.KeywordPaths == nil || len(porcupine.KeywordPaths) == 0 {
-		return fmt.Errorf("%s: No valid keywords were provided.", pvStatusToString(INVALID_ARGUMENT))
-	}
-
-	if porcupine.Sensitivities == nil {
-		porcupine.Sensitivities = make([]float32, len(porcupine.KeywordPaths))
-		for i := range porcupine.KeywordPaths {
-			porcupine.Sensitivities[i] = 0.5
-		}
-	}
-
-	if len(porcupine.KeywordPaths) != len(porcupine.Sensitivities) {
-		return fmt.Errorf("%s: Keyword array size (%d) is not the same size as sensitivities array (%d)",
-			pvStatusToString(INVALID_ARGUMENT), len(porcupine.KeywordPaths), len(porcupine.Sensitivities))
-	}
-
+func (porcupine *Porcupine) nativeInit() int {
 	var (
-		// modelPathC  = C.CString(porcupine.ModelPath)
+		modelPathC  = C.CString(porcupine.ModelPath)
 		numKeywords = len(porcupine.KeywordPaths)
 		keywordsC   = make([]*C.char, numKeywords)
 	)
@@ -239,120 +100,38 @@ func (porcupine *Porcupine) Init() (err error) {
 	}
 
 	var ret = C.pv_porcupine_init_wrapper(pv_porcupine_init_ptr,
-		C.CString(porcupine.ModelPath),
+		modelPathC,
 		(C.int32_t)(numKeywords),
 		(**C.char)(unsafe.Pointer(&keywordsC[0])),
 		(*C.float)(unsafe.Pointer(&porcupine.Sensitivities[0])),
 		(**C.pv_porcupine_t)(unsafe.Pointer(&porcupine.handle)))
 
-	if PvStatus(ret) != SUCCESS {
-		return fmt.Errorf(": Porcupine returned error %s", pvStatusToString(INVALID_ARGUMENT))
-	}
-
-	return nil
+	return int(ret)
 }
 
-// Releases resources acquired by Porcupine.
-func (porcupine *Porcupine) Delete() error {
-	if porcupine.handle == 0 {
-		return fmt.Errorf("Porcupine has not been initialized or has already been deleted.")
-	}
-
+func (porcupine *Porcupine) nativeDelete() {
 	C.pv_porcupine_delete_wrapper(pv_porcupine_delete_ptr,
 		(*C.pv_porcupine_t)(unsafe.Pointer(porcupine.handle)))
-	return nil
 }
 
-// Processes a frame of the incoming audio stream and emits the detection result.
-// Frame of audio The number of samples per frame can be attained by calling
-// `.FrameLength`. The incoming audio needs to have a sample rate equal to `.Sample` and be 16-bit
-// linearly-encoded. Porcupine operates on single-channel audio.
-// Returns a 0 based index if keyword was detected in frame. Returns -1 if no detection was made.
-func (porcupine *Porcupine) Process(pcm []int16) (int, error) {
-
-	if porcupine.handle == 0 {
-		return -1, fmt.Errorf("Porcupine has not been initialized or has been deleted.")
-	}
-
-	if len(pcm) != FrameLength {
-		return -1, fmt.Errorf("Input data frame is wrong size")
-	}
+func (porcupine *Porcupine) nativeProcess(pcm []int16) (int, int) {
 
 	var index int32
 	var ret = C.pv_porcupine_process_wrapper(pv_porcupine_process_ptr,
 		(*C.pv_porcupine_t)(unsafe.Pointer(porcupine.handle)),
 		(*C.int16_t)(unsafe.Pointer(&pcm[0])),
 		(*C.int32_t)(unsafe.Pointer(&index)))
-
-	if PvStatus(ret) != SUCCESS {
-		return -1, fmt.Errorf("Process audio frame failed with PvStatus: %d", ret)
-	}
-
-	return int(index), nil
+	return int(ret), int(index)
 }
 
-func getOS() string {
-	switch os := runtime.GOOS; os {
-	case "darwin":
-		return "mac"
-	case "linux":
-		return "linux"
-	case "windows":
-		return "windows"
-	default:
-		log.Fatalf("%s is not a supported OS", os)
-		return ""
-	}
+func nativeSampleRate() int {
+	return int(C.pv_sample_rate_wrapper(pv_sample_rate_ptr))
 }
 
-func extractDefaultModel() string {
-	modelPath := "embedded/lib/common/porcupine_params.pv"
-	return extractFile(modelPath, extractionDir)
+func nativeFrameLength() int {
+	return int(C.pv_porcupine_frame_length_wrapper(pv_porcupine_frame_length_ptr))
 }
 
-func extractKeywordFiles() map[string]string {
-	keywordDirPath := "embedded/resources/keyword_files/" + osName
-	keywordFiles, err := embeddedFS.ReadDir(keywordDirPath)
-	if err != nil {
-		log.Fatalf("%v", err)
-	}
-
-	extractedKeywords := make(map[string]string)
-	for _, keywordFile := range keywordFiles {
-		keywordPath := keywordDirPath + "/" + keywordFile.Name()
-		keywordName := strings.Split(keywordFile.Name(), "_")[0]
-		extractedKeywords[keywordName] = extractFile(keywordPath, extractionDir)
-	}
-	return extractedKeywords
-}
-
-func extractLib() string {
-	var libPath string
-	switch os := runtime.GOOS; os {
-	case "darwin":
-		libPath = fmt.Sprintf("embedded/lib/%s/x86_64/libpv_porcupine.dylib", osName)
-	case "linux":
-		libPath = fmt.Sprintf("embedded/lib/%s/x86_64/libpv_porcupine.so", osName)
-	case "windows":
-		libPath = fmt.Sprintf("embedded/lib/%s/amd64/libpv_porcupine.dll", osName)
-	default:
-		log.Fatalf("%s is not a supported OS", os)
-	}
-
-	return extractFile(libPath, extractionDir)
-}
-
-func extractFile(srcFile string, dstDir string) string {
-	bytes, readErr := embeddedFS.ReadFile(srcFile)
-	if readErr != nil {
-		log.Fatalf("%v", readErr)
-	}
-
-	extractedFilepath := filepath.Join(dstDir, srcFile)
-	os.MkdirAll(filepath.Dir(extractedFilepath), 0777)
-	writeErr := ioutil.WriteFile(extractedFilepath, bytes, 0777)
-	if writeErr != nil {
-		log.Fatalf("%v", writeErr)
-	}
-	return extractedFilepath
+func nativeVersion() string {
+	return int(C.pv_sample_rate_wrapper(pv_sample_rate_ptr))
 }
